발할라 연동 및 springboot 경로탐색 api, c-its 신호데이터 응답

+++ backend/src/main/java/kr/io/pacer/core/dto/external/SpatResponse.java
@@ -9,17 +9,17 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SpatResponse {
 
-    private String  itstId;
-    private Long    trsmUtcTime;
+    private String itstId;
+    private Long   trsmUtcTime;
 
-    private Integer ntPdsgRmdrCs;
-    private Integer etPdsgRmdrCs;
-    private Integer stPdsgRmdrCs;
-    private Integer wtPdsgRmdrCs;
-    private Integer nePdsgRmdrCs;
-    private Integer sePdsgRmdrCs;
-    private Integer swPdsgRmdrCs;
-    private Integer nwPdsgRmdrCs;
+    private Double ntPdsgRmdrCs;
+    private Double etPdsgRmdrCs;
+    private Double stPdsgRmdrCs;
+    private Double wtPdsgRmdrCs;
+    private Double nePdsgRmdrCs;
+    private Double sePdsgRmdrCs;
+    private Double swPdsgRmdrCs;
+    private Double nwPdsgRmdrCs;
 
     public int getItstIdAsInt() {
         return Integer.parseInt(itstId);
