+++ backend/src/main/java/kr/io/pacer/core/dto/response/RouteResponse.java
@@ -30,14 +30,14 @@ public class RouteResponse {
         private String  name;
         private double  lat;
         private double  lng;
-        private Integer ntPdsgRmdrCs;
-        private Integer etPdsgRmdrCs;
-        private Integer stPdsgRmdrCs;
-        private Integer wtPdsgRmdrCs;
-        private Integer nePdsgRmdrCs;
-        private Integer sePdsgRmdrCs;
-        private Integer swPdsgRmdrCs;
-        private Integer nwPdsgRmdrCs;
+        private Double ntPdsgRmdrCs;
+        private Double etPdsgRmdrCs;
+        private Double stPdsgRmdrCs;
+        private Double wtPdsgRmdrCs;
+        private Double nePdsgRmdrCs;
+        private Double sePdsgRmdrCs;
+        private Double swPdsgRmdrCs;
+        private Double nwPdsgRmdrCs;
     }
 
     @Getter
